Refactor submodule hierarchy
