+++ core/markets/__init__.py
@@ -6,3 +6,5 @@ Notes:
     `SimulatedMarket` is an ideal market (accepts all trades) that is used for general backtesting. It should be
     improved *by not accepting all orders* to more accurately model real world `Strategy` performance.
 """
+from core.markets.GeminiMarket import GeminiMarket
+from core.markets.SimulatedMarket import SimulatedMarket
