+++ models/indicators/__init__.py
@@ -1,0 +1,3 @@
+from models.indicators.BBANDSRow import BBANDSRow
+from models.indicators.MACDRow import MACDRow
+from models.indicators.STOCHRSIRow import STOCHRSIRow
