+++ primitives/__init__.py
@@ -1,0 +1,2 @@
+from primitives.signals import Signal, Side
+from primitives.trends import MarketTrend, TrendMovement
