+++ requirements.txt
@@ -5,7 +5,7 @@ requests~=2.28.1
 pandas==1.5.2
 jupyter
 TA-lib==0.4.25
-apscheduler==3.9.1
+apscheduler==3.9.1.post1
 pyyaml~=6.0
 urllib3~=1.26.2
 pytz==2022.6
