+++ strategies/__init__.py
@@ -8,3 +8,7 @@
     To determine the fitness and performance of the trading strategy, reporting functions can the total amount of
     assets and fiat accrued. This can be used in active implementations as well as during backtesting.
 """
+from strategies.strategy import Strategy
+from strategies.OscillatingStrategy import OscillatingStrategy
+from strategies.StaticAlternatingStrategy import StaticAlternatingStrategy
+from strategies.ThreeProngAlt import ThreeProngAlt
